initial

+++ src/main/java/com/lesson/calculator/CalcService.java
@@ -4,5 +4,5 @@ public interface CalcService {
     int plus(int a, int b);
     int minus(int a, int b);
     int multiply(int a, int b);
-    int divide(int a, int b);
+    double divide(double a, double b);
 }
